fix: address UI audit gaps across storefront
Fixes found during a full UI audit:
- remove dead /node_modules fontawesome stylesheet link and stray
  server-only mongoose import in a client component
- add app/error.js, app/loading.js, app/not-found.js route boundaries
- add per-page metadata (product, category, search, about, contact,
  products listing) by splitting each into a server page.js that
  exports metadata/generateMetadata and a client component it renders
- aria-label pass on icon-only buttons in the core shopping flow
  (wishlist, cart, pagination), plus fix a hamburger menu aria-label
  that was reading isScrolled instead of isMenuOpen
- wire up the newsletter "Unsubscribe anytime" link to a real
  /unsubscribe page backed by a new DELETE /api/newsletter handler
- add vitest + smoke tests for the newsletter subscribe/unsubscribe API
