+++ app/(main)/category/[slug]/CategoryPage.jsx
@@ -1,0 +1,357 @@
+"use client";
+
+import React, { useState, useEffect } from "react";
+import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { toast } from "react-toastify";
+import Link from "next/link";
+import { getProducts } from "@/actions/products";
+import { getWishlist, updateWishlist } from "@/actions/wishlist";
+import { addToCart, getCart } from "@/actions/cart-utils";
+import { useSession } from "next-auth/react";
+import { useRouter } from "next/navigation";
+import ProductCard from "@/components/ProductCard";
+
+const CategoryPage = ({ params }) => {
+  const { slug } = params;
+  const queryClient = useQueryClient();
+  const { data: session, status } = useSession();
+  const router = useRouter();
+
+  const [currentPage, setCurrentPage] = useState(1);
+  const itemsPerPage = 12;
+
+  const decodedSlug = decodeURIComponent(slug);
+
+  const formattedCategoryName = decodedSlug
+    .split("-")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
+    .join(" ");
+
+  // Fetch products
+  const { data, error, isLoading } = useQuery({
+    queryKey: ["categoryProducts", slug],
+    queryFn: () => getProducts({ category: decodedSlug }),
+    onError: (error) => {
+      toast.error(`Error fetching products: ${error.message}`, {
+        position: "top-right",
+        autoClose: 3000,
+      });
+    },
+  });
+
+  // Fetch wishlist
+  const { data: wishlistData, error: wishlistError, isLoading: wishlistLoading } = useQuery({
+    queryKey: ["wishlist"],
+    queryFn: getWishlist,
+    enabled: status === "authenticated",
+    onError: (error) => {
+      if (!error.message.includes("Unauthorized")) {
+        toast.error(`Error fetching wishlist: ${error.message}`, {
+          position: "top-right",
+          autoClose: 3000,
+        });
+      }
+    },
+  });
+
+  // Fetch cart
+  const { data: cartData, error: cartError, isLoading: cartLoading } = useQuery({
+    queryKey: ["cart"],
+    queryFn: getCart,
+    enabled: status === "authenticated",
+    onError: (error) => {
+      if (error.message.includes("Unauthorized")) {
+        toast.error("Please log in to view your cart.", {
+          position: "top-right",
+          autoClose: 3000,
+        });
+        setTimeout(() => router.push("/login"), 3000);
+      } else {
+        toast.error(`Error fetching cart: ${error.message}`, {
+          position: "top-right",
+          autoClose: 3000,
+        });
+      }
+    },
+  });
+
+  // Mutation to add/remove product from wishlist
+  const wishlistMutation = useMutation({
+    mutationFn: ({ productId, action }) => updateWishlist(productId, action),
+    onSuccess: (data) => {
+      queryClient.invalidateQueries(["wishlist"]);
+      toast.success(data.message, {
+        position: "top-right",
+        autoClose: 3000,
+      });
+    },
+    onError: (error) => {
+      if (error.message.includes("Unauthorized")) {
+        toast.error("Please log in to manage your wishlist.", {
+          position: "top-right",
+          autoClose: 3000,
+        });
+        setTimeout(() => router.push("/login"), 3000);
+      } else {
+        toast.error(`Error: ${error.message}`, {
+          position: "top-right",
+          autoClose: 3000,
+        });
+      }
+    },
+  });
+
+  // Cart mutation
+  const cartMutation = useMutation({
+    mutationFn: ({ productId, quantity }) => addToCart(productId, quantity),
+    onSuccess: (data) => {
+      queryClient.invalidateQueries(["cart"]);
+      toast.success("Product added to cart successfully!", {
+        position: "top-right",
+        autoClose: 3000,
+      });
+    },
+    onError: (error) => {
+      if (error.message.includes("Unauthorized")) {
+        toast.error("Please log in to add to cart.", {
+          position: "top-right",
+          autoClose: 3000,
+        });
+        setTimeout(() => router.push("/login"), 3000);
+      } else {
+        toast.error(`Error: ${error.message}`, {
+          position: "top-right",
+          autoClose: 3000,
+        });
+      }
+    },
+  });
+
+  // Check if a product is in the cart
+  const isInCart = (productId) => {
+    return cartData?.cart?.items?.some(
+      (item) => item.product._id.toString() === productId
+    ) || false;
+  };
+
+  // Handle adding/removing from wishlist
+  const handleWishlistToggle = (productId, e) => {
+    e.preventDefault();
+    if (status !== "authenticated") {
+      toast.error("Please log in to manage your wishlist.", {
+        position: "top-right",
+        autoClose: 3000,
+      });
+      setTimeout(() => router.push("/login"), 3000);
+      return;
+    }
+    if (wishlistLoading || wishlistMutation.isLoading) return;
+    const action = wishlistData?.wishlist?.some((item) => item._id === productId) ? "remove" : "add";
+    wishlistMutation.mutate({ productId, action });
+  };
+
+  // Handle add to cart or view cart
+  const handleCartAction = (productId, e) => {
+    e.preventDefault();
+    if (status !== "authenticated") {
+      toast.error("Please log in to add to cart.", {
+        position: "top-right",
+        autoClose: 3000,
+      });
+      setTimeout(() => router.push("/login"), 3000);
+      return;
+    }
+
+    if (isInCart(productId)) {
+      router.push("/cart");
+    } else {
+      if (cartMutation.isLoading) return;
+      cartMutation.mutate({ productId, quantity: 1 });
+    }
+  };
+
+  // Reset to first page when category changes
+  useEffect(() => {
+    setCurrentPage(1);
+  }, [slug]);
+
+  if (isLoading) {
+    return (
+      <div className="flex justify-center items-center min-h-[500px]">
+        <div className="animate-spin rounded-full h-20 w-20 border-b-2 border-gray-900"></div>
+      </div>
+    );
+  }
+
+  if (error) {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <p>Failed to load products. Please try again later.</p>
+      </div>
+    );
+  }
+
+  const products = data?.products || [];
+
+  // Pagination logic
+  const totalItems = products.length;
+  const totalPages = Math.max(1, Math.ceil(totalItems / itemsPerPage));
+  
+  // Ensure current page is within valid range
+  const validCurrentPage = Math.min(Math.max(1, currentPage), totalPages);
+  if (validCurrentPage !== currentPage) {
+    setCurrentPage(validCurrentPage);
+  }
+  
+  const startIndex = (validCurrentPage - 1) * itemsPerPage;
+  const endIndex = Math.min(startIndex + itemsPerPage, totalItems);
+  const paginatedProducts = products.slice(startIndex, endIndex);
+
+  // Handle page change
+  const handlePageChange = (page) => {
+    if (page < 1 || page > totalPages) return;
+    setCurrentPage(page);
+    window.scrollTo({ top: 0, behavior: "smooth" });
+  };
+
+  // Generate page numbers for pagination display
+  const getPageNumbers = () => {
+    const pageNumbers = [];
+    const maxVisiblePages = 5;
+    
+    if (totalPages <= maxVisiblePages) {
+      for (let i = 1; i <= totalPages; i++) {
+        pageNumbers.push(i);
+      }
+    } else {
+      if (validCurrentPage <= 3) {
+        for (let i = 1; i <= 4; i++) {
+          pageNumbers.push(i);
+        }
+        pageNumbers.push('...');
+        pageNumbers.push(totalPages);
+      } else if (validCurrentPage >= totalPages - 2) {
+        pageNumbers.push(1);
+        pageNumbers.push('...');
+        for (let i = totalPages - 3; i <= totalPages; i++) {
+          pageNumbers.push(i);
+        }
+      } else {
+        pageNumbers.push(1);
+        pageNumbers.push('...');
+        for (let i = validCurrentPage - 1; i <= validCurrentPage + 1; i++) {
+          pageNumbers.push(i);
+        }
+        pageNumbers.push('...');
+        pageNumbers.push(totalPages);
+      }
+    }
+    
+    return pageNumbers;
+  };
+
+  return (
+    <>
+      <div className="min-h-screen py-10 bg-white">
+        <div className="mx-auto max-w-7xl px-4">
+          <h1 className="text-2xl font-semibold text-gray-800 mb-8 text-center uppercase">
+            {formattedCategoryName}
+          </h1>
+          {products.length === 0 ? (
+            <div className="flex flex-col items-center justify-center h-full">
+              <p className="text-center text-gray-600 text-lg mb-4">
+                No products found in this category.
+              </p>
+              <Link
+                href="/products"
+                className="px-6 py-2 text-center text-sm text-white bg-primary border border-primary rounded hover:bg-transparent hover:text-primary transition uppercase font-roboto font-medium"
+              >
+                Browse All Products
+              </Link>
+            </div>
+          ) : (
+            <>
+              <div className="mb-4 text-gray-600">
+                Showing {startIndex + 1}-{endIndex} of {totalItems} products
+              </div>
+              
+              <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-4 gap-6">
+                {paginatedProducts.map((product) => (
+                  <ProductCard
+                    key={product._id}
+                    product={product}
+                    wishlistData={wishlistData}
+                    wishlistLoading={wishlistLoading}
+                    mutation={wishlistMutation}
+                    handleWishlistToggle={handleWishlistToggle}
+                    cartData={cartData}
+                    cartLoading={cartLoading}
+                    cartMutation={cartMutation}
+                    handleCartAction={handleCartAction}
+                    isInCart={isInCart}
+                  />
+                ))}
+              </div>
+
+              {/* Pagination Controls */}
+              {totalPages > 1 && (
+                <div className="flex justify-center mt-8">
+                  <nav className="flex items-center gap-2">
+                    <button
+                      onClick={() => handlePageChange(validCurrentPage - 1)}
+                      disabled={validCurrentPage === 1}
+                      className={`px-4 py-2 text-sm font-medium rounded ${
+                        validCurrentPage === 1
+                          ? "bg-gray-200 text-gray-500 cursor-not-allowed"
+                          : "bg-primary text-white hover:bg-primary-dark"
+                      }`}
+                    >
+                      Previous
+                    </button>
+
+                    {getPageNumbers().map((page, index) =>
+                      page === "..." ? (
+                        <span
+                          key={`ellipsis-${index}`}
+                          className="px-4 py-2 text-sm font-medium"
+                        >
+                          {page}
+                        </span>
+                      ) : (
+                        <button
+                          key={`page-${page}`}
+                          onClick={() => handlePageChange(page)}
+                          className={`px-4 py-2 text-sm font-medium rounded ${
+                            validCurrentPage === page
+                              ? "bg-primary text-white"
+                              : "bg-gray-200 text-gray-700 hover:bg-gray-300"
+                          }`}
+                        >
+                          {page}
+                        </button>
+                      )
+                    )}
+
+                    <button
+                      onClick={() => handlePageChange(validCurrentPage + 1)}
+                      disabled={validCurrentPage === totalPages}
+                      className={`px-4 py-2 text-sm font-medium rounded ${
+                        validCurrentPage === totalPages
+                          ? "bg-gray-200 text-gray-500 cursor-not-allowed"
+                          : "bg-primary text-white hover:bg-primary-dark"
+                      }`}
+                    >
+                      Next
+                    </button>
+                  </nav>
+                </div>
+              )}
+            </>
+          )}
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default CategoryPage;
