+++ app/layout.js
@@ -19,8 +19,12 @@ const geistMono = localFont({
 });
 
 export const metadata = {
-  title: "SwiftCart",
-  description: "Generated by create next app",
+  title: {
+    default: "SwiftCart",
+    template: "%s | SwiftCart",
+  },
+  description:
+    "Shop the latest home decor, furniture, and lifestyle products at SwiftCart.",
 };
 
 export default function RootLayout({ children }) {
@@ -36,10 +40,6 @@ export default function RootLayout({ children }) {
         <link
           href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&family=Roboto:wght@400;500;700&display=swap"
           rel="stylesheet"
-        />
-        <link
-          rel="stylesheet"
-          href="/node_modules/@fortawesome/fontawesome-free/css/all.min.css"
         />
       </head>
       <body
