+++ components/ProductsCard.jsx
@@ -161,6 +161,7 @@ const ProductCard = ({ product, wishlistData, cartData, queryClient }) => {
             href={`/products/${product._id}`}
             className="text-white text-lg w-9 h-8 rounded-full bg-primary flex items-center justify-center hover:bg-gray-800 transition"
             title="View Product"
+            aria-label="View Product"
           >
             <FaEye />
           </Link>
@@ -173,6 +174,7 @@ const ProductCard = ({ product, wishlistData, cartData, queryClient }) => {
                 : "bg-primary hover:bg-gray-800"
             } ${mutation.isPending ? "opacity-50 cursor-not-allowed" : ""}`}
             title={isInWishlist ? "Remove from Wishlist" : "Add to Wishlist"}
+            aria-label={isInWishlist ? "Remove from Wishlist" : "Add to Wishlist"}
           >
             {mutation.isPending ? (
               <svg
