+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test": "vitest run"
   },
   "dependencies": {
     "@auth/mongodb-adapter": "^3.8.0",
@@ -32,6 +33,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.24",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "^3.4.1",
+    "vitest": "^4.1.10"
   }
 }
